[checker] Move toplevel name collision checkers to its own file
